requirements.txt
openai-agents>=0.8.0
httpx>=0.27.0
playwright>=1.49.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
apscheduler>=3.10.0
